requirements.txt
openai
pinecone-client
python-dotenv
